+++ extensions/rss-reader/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [Add Favicons and Detail View] - 2022-09-16
 
-- Add favicons to the list of feeds
-- Switch to story list when adding a new feed
-- Add feed title as story subtitle in the story list
+- Add favicons and subtitles to the story list view
+- Switch to story list view when adding a new feed
+- Add feed title as story subtitle in the story list view
+- add confirmation when deleting a feed
 - Add detail view to read story from raycast
+- add dropdown to select feed in the story list
+- add confirmation alert when running a destructive action